backend/requirements.txt
# FastAPI and Serverless
fastapi==0.109.0
uvicorn==0.27.0
mangum==0.17.0
pydantic==2.5.3

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
alembic==1.13.1

# RAG System
qdrant-client==1.7.0
openai==1.10.0
langchain==0.1.0
langchain-community==0.0.20
sentence-transformers==2.3.1

# Testing
pytest==8.0.0
pytest-asyncio==0.23.3
httpx==0.26.0
pytest-cov==4.1.0

# Utilities
python-dotenv==1.0.0
